Organize React files and establish routing
