+++ package.json
@@ -6,6 +6,7 @@
     "antd": "^3.20.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
+    "react-router-dom": "^5.0.1",
     "react-scripts": "3.0.1"
   },
   "scripts": {
